// ui/src/pages/Trace/model/trace-dag/types/TDagNode.tsx
// Copyright (c) 2020 The Jaeger Authors.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
// http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

import { NodeID } from './index';

type TDagNode<TData extends { [k: string]: unknown } = {}> = TData & {
  parentID: NodeID | null;
  id: NodeID;
  children: Set<NodeID>;
};

// eslint-disable-next-line no-undef
export default TDagNode;